requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
google-genai==1.68.0
supabase==2.11.0
websockets==15.0.1
python-dotenv==1.0.0
Flask-Limiter==3.5.0
python-dateutil==2.9.0